+++ plume/src/main/kotlin/io/github/plume/oss/passes/graph/CGPass.kt
@@ -15,14 +15,15 @@
  */
 package io.github.plume.oss.passes.graph
 
-import io.github.plume.oss.GlobalCache
+import io.github.plume.oss.cache.GlobalCache
+import io.github.plume.oss.domain.model.DeltaGraph
 import io.github.plume.oss.drivers.IDriver
-import io.github.plume.oss.options.ExtractorOptions
 import io.github.plume.oss.passes.IUnitGraphPass
 import io.shiftleft.codepropertygraph.generated.EdgeTypes.CALL
 import io.shiftleft.codepropertygraph.generated.NodeKeyNames.FULL_NAME
 import io.shiftleft.codepropertygraph.generated.NodeTypes.METHOD
 import io.shiftleft.codepropertygraph.generated.nodes.NewMethodBuilder
+import io.shiftleft.codepropertygraph.generated.nodes.NewNodeBuilder
 import org.apache.logging.log4j.LogManager
 import soot.Scene
 import soot.Unit
@@ -36,26 +37,23 @@ import soot.toolkits.graph.BriefUnitGraph
 /**
  * The [IUnitGraphPass] that constructs the interprocedural call edges.
  *
+ * @param g The driver to build the call edges with.
  * @param driver The driver to build the call edges with.
  */
-class CGPass(private val driver: IDriver) : IUnitGraphPass {
+class CGPass(private val g: BriefUnitGraph, private val driver: IDriver) : IUnitGraphPass {
+
     private val logger = LogManager.getLogger(CGPass::javaClass)
-    private lateinit var g: BriefUnitGraph
+    private val builder = DeltaGraph.Builder()
 
-    override fun runPass(gs: List<BriefUnitGraph>) =
-        if (ExtractorOptions.callGraphAlg != ExtractorOptions.CallGraphAlg.NONE) gs.map(::runPassOnGraph)
-        else gs
-
-    private fun runPassOnGraph(g: BriefUnitGraph): BriefUnitGraph {
+    override fun runPass(): DeltaGraph {
         val mtd = g.body.method
-        logger.debug("Building call graph edges for ${mtd.declaration}")
-        this.g = g
+        logger.debug("Building call graph edges for ${mtd.declaringClass.name}:${mtd.name}")
         // If this was an updated method, connect call graphs
         GlobalCache.getMethodCache(mtd).filterIsInstance<NewMethodBuilder>()
             .firstOrNull()?.let { reconnectPriorCallGraphEdges(it) }
-        // Connect all units to their successors
+        // Connect all calls to their methods
         this.g.body.units.filterNot { it is IdentityStmt }.forEach(this::projectUnit)
-        return g
+        return builder.build()
     }
 
     private fun projectUnit(unit: Unit) {
@@ -73,8 +71,8 @@ class CGPass(private val driver: IDriver) : IUnitGraphPass {
                 GlobalCache.getMethodCache(e.tgt.method())
                     .filterIsInstance<NewMethodBuilder>()
                     .firstOrNull()?.let { tgtPlumeVertex ->
-                        runCatching { driver.addEdge(callV, tgtPlumeVertex, CALL); foundAndConnectedCallTgt = true }
-                            .onFailure { e -> logger.warn(e.message) }
+                        builder.addEdge(callV, tgtPlumeVertex, CALL)
+                        foundAndConnectedCallTgt = true
                     }
             }
             // If call graph analysis fails because there is no main method, we will need to figure out call edges ourselves
@@ -83,23 +81,24 @@ class CGPass(private val driver: IDriver) : IUnitGraphPass {
             if (!foundAndConnectedCallTgt) {
                 val v = callV.build()
                 if (v.methodFullName().length > 1) {
-                    driver.getVerticesByProperty(FULL_NAME, v.methodFullName(), METHOD).firstOrNull()
-                        ?.let { mtdV -> driver.addEdge(callV, mtdV, CALL) }
+                    getMethodHead(v.methodFullName())?.let { mtdV -> builder.addEdge(callV, mtdV, CALL) }
                 }
             }
         }
     }
+
+    private fun getMethodHead(fullName: String): NewNodeBuilder? =
+        GlobalCache.getMethod(fullName)
+            ?: driver.getVerticesByProperty(FULL_NAME, fullName, METHOD).firstOrNull()
+                ?.apply { GlobalCache.addMethod(this as NewMethodBuilder) }
+
 
     private fun reconnectPriorCallGraphEdges(mtdV: NewMethodBuilder) {
         val mtd = mtdV.build()
         GlobalCache.getCallEdgeIn(mtd.fullName())?.let { incomingVs ->
             if (incomingVs.isNotEmpty()) {
                 logger.debug("Saved call graph edges found - reconnecting incoming call graph edges")
-                incomingVs.forEach { inV ->
-                    runCatching {
-                        driver.addEdge(inV, mtdV, CALL)
-                    }.onFailure { e -> logger.warn(e.message) }
-                }
+                incomingVs.forEach { inV -> builder.addEdge(inV, mtdV, CALL) }
             } else {
                 logger.debug("No previous call graph edges were found")
             }